Title: 두 수의 합, Time: 103ms, Memory: 8MB, Status: Wrong Answer - CodeTree

+++ 240422/두 수의 합/sum-of-two-num.java
@@ -5,23 +5,23 @@ public class Main {
     static BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
     static StringTokenizer st;
     static int n, k;
-    static int[] nums;
-    static HashMap<Integer, Integer> hm = new HashMap<>();
+    static long[] nums;
+    static HashMap<Long, Integer> hm = new HashMap<>();
 
     public static void main(String[] args) throws IOException {
         st = new StringTokenizer(br.readLine());
         n = Integer.parseInt(st.nextToken());
         k = Integer.parseInt(st.nextToken());
 
-        nums = new int[n];
+        nums = new long[n];
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < n; i++) {
-            nums[i] = Integer.parseInt(st.nextToken());
+            nums[i] = Long.parseLong(st.nextToken());
         }
 
         for (int i = 0; i < n - 1; i++) {
             for (int j = i + 1; j < n; j++) {
-                int sum = nums[i] + nums[j];
+                long sum = nums[i] + nums[j];
                 hm.put(sum, hm.getOrDefault(sum, 0) + 1);
             }
         }
